Add support for connecting via an ssh key
For this version of the interface, an ssh key should be specified at the
time of making the connection.  The other way we could have handled
things would have been to add a method to set the ssh key and, after
that, use the normal connect with a null password.  I'm open to
suggestions of going that way.